Drew - 問題回報可以存進資料庫了!!! 感謝漢彬，忠和，柏翔 :)))))))))

+++ src/main/java/foodelicious/compbackend/repository/CBKProblemRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import foodelicious.compbackend.model.ProblemsBean;
 
 @Repository
-public interface CBKProblemRepository extends JpaRepository<ProblemsBean, Long> {
+public interface CBKProblemRepository extends JpaRepository<ProblemsBean, Integer> {
 
 }
